refactor: move lsp helper functions to common module
