+++ wdl-analysis/src/handlers/common.rs
@@ -1,0 +1,7 @@
+//! Common utilities for LSP handlers.
+
+mod docs;
+mod position;
+
+pub use docs::*;
+pub use position::*;
